chromium/third_party/blink/web_tests/virtual/out-of-blink-cspee/external/wpt/content-security-policy/embedded-enforcement/subsumption_algorithm-self-expected.txt

This is a testharness.js-based test.
PASS 'self' keywords should match.
PASS Returned CSP does not have to specify 'self'.
FAIL Returned CSP must not allow 'self' if required CSP does not. assert_equals: expected (undefined) undefined but got (boolean) true
PASS Returned 'self' should match to an origin's url.
PASS Required 'self' should match to a origin's url.
PASS Required 'self' should subsume a more secure version of origin's url.
FAIL Returned 'self' should not be subsumed by a more secure version of origin's url. assert_unreached: No message should be sent from the frame. Reached unreachable code
Harness: the test ran to completion.
